Fix -Wconversion warnings where possible.
In order to detect "int foo = myQReal" mistakes, one needs to compile with -Wconversion.
However that flag triggers many warnings inside Qt. This commit fixes many of them, like
qchar.h:295: warning: conversion to 'ushort' from 'unsigned int' may alter its value
qglobal.h:1026: warning: conversion to 'qreal' from 'qint64' may alter its value
qbytearray.h:441: warning: conversion to 'char' from 'int' may alter its value
Other warnings remain (such as those coming from bitfields)

Merge-request: 1460

+++ src/corelib/global/qendian.h
@@ -149,9 +149,9 @@ template <> inline quint32 qFromLittleEndian<quint32>(const uchar *src)
 
 template <> inline quint16 qFromLittleEndian<quint16>(const uchar *src)
 {
-    return 0
-        | src[0]
-        | src[1] * 0x0100;
+    return quint16(0
+                   | src[0]
+                   | src[1] * 0x0100);
 }
 
 // signed specializations
@@ -241,9 +241,9 @@ inline quint32 qFromBigEndian<quint32>(const uchar *src)
 template<>
 inline quint16 qFromBigEndian<quint16>(const uchar *src)
 {
-    return 0
-        | src[1]
-        | src[0] * quint16(0x0100);
+    return quint16( 0
+                    | src[1]
+                    | src[0] * quint16(0x0100));
 }
 
 
@@ -288,9 +288,9 @@ template <> inline quint32 qbswap<quint32>(quint32 source)
 
 template <> inline quint16 qbswap<quint16>(quint16 source)
 {
-    return 0
-        | ((source & 0x00ff) << 8)
-        | ((source & 0xff00) >> 8);
+    return quint16( 0
+                    | ((source & 0x00ff) << 8)
+                    | ((source & 0xff00) >> 8) );
 }
 
 // signed specializations
